+++ src/main/java/com/revolvingmadness/sculk/language/builtins/classes/types/ListType.java
@@ -7,6 +7,7 @@ import com.revolvingmadness.sculk.language.builtins.classes.BuiltinType;
 import com.revolvingmadness.sculk.language.builtins.classes.instances.BooleanInstance;
 import com.revolvingmadness.sculk.language.builtins.classes.instances.IntegerInstance;
 import com.revolvingmadness.sculk.language.builtins.classes.instances.ListInstance;
+import com.revolvingmadness.sculk.language.builtins.classes.instances.NullInstance;
 import com.revolvingmadness.sculk.language.interpreter.Interpreter;
 import com.revolvingmadness.sculk.language.lexer.TokenType;
 
@@ -17,6 +18,7 @@ public class ListType extends BuiltinType {
         super("List");
         this.typeVariableScope.declare(List.of(TokenType.CONST), "length", new Length());
         this.typeVariableScope.declare(List.of(TokenType.CONST), "contains", new Contains());
+        this.typeVariableScope.declare(List.of(TokenType.CONST), "append", new Append());
     }
 
     @Override
@@ -44,6 +46,21 @@ public class ListType extends BuiltinType {
             BuiltinClass other = arguments.get(0);
 
             return new BooleanInstance(this.boundClass.toList().contains(other));
+        }
+    }
+
+    private static class Append extends BuiltinMethod {
+        @Override
+        public BuiltinClass call(Interpreter interpreter, List<BuiltinClass> arguments) {
+            if (arguments.size() != 1) {
+                throw ErrorHolder.invalidArgumentCount("append", 1, arguments.size());
+            }
+
+            BuiltinClass object = arguments.get(0);
+
+            this.boundClass.toList().add(object);
+
+            return new NullInstance();
         }
     }